Switched back from mini.comment to Comment.nvim
mini.comment is very simple and does not set commentstring by itself. So I need to use something else to set it.
This is where nvim-ts-context-commentstring comes in. It uses treesitter in order to set the commentstring.
However, I've been having issues with treesitter at work so have disabled it there. Which means mini.comment
doesn't work very well so I've been using Comment.nvim when treesitter is disabled and mini.comment when it is.
This is too cumbersome so in this changelist I just switched to Comment.nvim entirely

as there were too many hoops to jump through to use it only when tree-sitter was enabled